requirements.txt
anthropic==0.42.0
Flask==3.1.0
google_auth_oauthlib==1.2.1
langchain==0.3.13
langchain_community==0.3.13
langchain_core==0.3.28
langchain_google_genai==2.0.7
langchain_huggingface==0.1.2
langchain_mongodb==0.3.0
langchain_text_splitters==0.3.4
llamaapi==0.1.36
numpy==2.2.1
ollama==0.4.4
openai==1.58.1
protobuf==5.29.2
pymongo==4.10.1
python-dotenv==1.0.1
Requests==2.32.3
sentence_transformers==3.0.1
